+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+#1.2.4
+- Added separate configs for icons and labels opacity
+- Added 'Circle location' config (always under cursor / always corner / Under cursor + in corner when no space)
+- Changed some default configs
+- Small options page improvements
+
 #1.2.3
 - Fixed circle transition not working on some websites